ValuesUtils
Add a quick note about limitations of using remote autocompletion with
'values from concept' and displaytitle. In order to run a query on a
concept and use the displaytitle to search on, 'Display title of'
needs to be available as a property for that page. The user will have
to add the DISPLAYTITLE magic word.

Select2:

- Fix for double-click events. Since it is only labels not values that
can be extracted from token elements, the value corresponding to a label
needs to be looked up first; when this is done, remove clicked from
selected values and feed the updated array to $input (largely restoring
the previous method used in PF).

- Fix for duplicate tokens - element.on("select2:select"...). For
reasons that are yet unclear, double-clicking a token may result in an
option being appended to one of two select/options elements (the hidden
one) even if one with the same value exists. As a result,
double-clicking and confirming the same value from the dropdown
list can lead to seeing two identical tokens being rendered in the
same container. This is a bit of a hack in that it looks for the
attribute data-select2-tag="true" (not attested otherwise) so that it
can remove the relevant option.

 - Adopt value/label mechanism with local autocompletion